refactor: move blueprint/pipeline swagger docs into plugin folder
